+++ libraries/PCF8575/CHANGELOG.md
@@ -7,8 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
 ## [0.1.10] - 2023-09-24
-- fix Wire1 support for ESP32 (again)
+- refactor API, begin()
+- update examples
+- update readme.md
 
+----
+
+## [0.1.10] - 2023-09-24
+- fix Wire1 support for ESP32 (again)
 
 ## [0.1.9] - 2023-09-23
 - add Wire1 support for ESP32